+++ FishExpenseManager/mobile/lib/features/customers/domain/repositories/customer_repository.dart
@@ -3,6 +3,7 @@ import '../entities/customer_entity.dart';
 abstract class CustomerRepository {
   Future<List<CustomerEntity>> getAllCustomers();
   Stream<List<CustomerEntity>> watchAllCustomers();
+  Stream<CustomerEntity?> watchCustomerById(int id);
   Future<void> saveCustomer(CustomerEntity customer);
   Future<int> addCustomer(CustomerEntity customer);
   Future<void> deleteCustomer(int id);